// src/app/maps/interfaces/places.ts
// Generated by https://quicktype.io

export interface PlacesResponse {
  type:        string;
  query:       string[];
  features:    Feature[];
  attribution: string;
}

export interface Feature {
  id:                   string;
  type:                 FeatureType;
  place_type:           PlaceType[];
  relevance:            number;
  properties:           Properties;
  text:                 string;
  place_name:           string;
  matching_text?:       string;
  matching_place_name?: string;
  bbox:                 number[];
  center:               number[];
  geometry:             Geometry;
  context:              Context[];
}

export interface Context {
  id:         string;
  short_code: string;
  wikidata:   string;
  text:       string;
}

export interface Geometry {
  type:        GeometryType;
  coordinates: number[];
}

export enum GeometryType {
  Point = "Point",
}

export enum PlaceType {
  Place = "place",
}

export interface Properties {
  wikidata: string;
}

export enum FeatureType {
  Feature = "Feature",
}